+++ cas-server-support-jpa-ticket-registry/build.gradle
@@ -42,5 +42,6 @@ dependencies {
   testCompile project(path: ":cas-server-core-authentication", configuration: "tests")
   testCompile project(':cas-server-core-logout')
   testCompile project(':cas-server-core-services')
+  testCompile project(':cas-server-core')
 
 }
