+++ cas-server-support-jpa-ticket-registry/src/test/resources/jpaSpringContext.xml
@@ -24,7 +24,7 @@
   <context:component-scan base-package="org.jasig.cas" />
   <context:annotation-config/>
 
-  <import resource="classpath:META-INF/spring/jpa-ticket-reg-context.xml" />
+  <import resource="classpath:jpaTicketRegistryContext.xml" />
 
   <bean
       id="dataSourceTicket"
